Custiomize code and build components
